refactor(form/inputs): keep unstable onChange prop out of hook deps
Use `onChange` from `useFormCallbacks` hook instead of passing it around as props to the PT-input's hook deps
and internal components.

It's a unstable prop, and is causing a total re-render of all the PT-input's blocks at every time the value changes.
The onChange props seems to be tied to the value (sometimes it produces unset patches other times not), so we
definately need to use the latest version always.