Add support for block form components in config and plugins (#4327)
* feature(core/form): add global render support for block components

This will allow writing global changes with the studio config or plugins,
to how blocks, inlineBlocks and annotations are rendered in the studio.

Brings parity with how array item is working.

* dev(test-studio): update schemas with block form components

Include block, inlineBlock and annotation custom rendering example.

* test(form): update tests for renderArrayOfObjects/renderObject input

Update according to recent changes with block form components.

* fix(form/inputs): include form for every PT member item

Include the form for every member, not just the opened one.

* refactor(form/inputs): remove merge leftovers resolving conflict

* refactor(inputs/form): restore now stable props.path

Path from props here was previosly unstable, but stabilized via https://github.com/sanity-io/sanity/pull/4335

Remove memoization of it.

* refactor(form/inputs): keep unstable onChange prop out of hook deps

Use `onChange` from `useFormCallbacks` hook instead of passing it around as props to the PT-input's hook deps
and internal components.

It's a unstable prop, and is causing a total re-render of all the PT-input's blocks at every time the value changes.
The onChange props seems to be tied to the value (sometimes it produces unset patches other times not), so we
definately need to use the latest version always.

* fix(form/inputs): disable warning for empty useMemo dep.

Still need to figure out the unstability of this prop, but keep caching for now.

es-lint started hard-failing on these now.

* refactor(core/form/inputs): remove unstable prop workaround

* refactor(core/form/inputs): make new block render cb-props optional

In order to not introduce breaking changes, we need to keep the new
block related render callbacks optional.

Several of the existing ones should maybe also be optional,
but for now it has to be like this.

* chore(core/form/inputs): remove rebase leftovers

* refactor(core/form/types): children is only a ReactNode

Align with itemProps, and this is always returned now.

* chore(core/form/inputs): fix rebase leftovers